+++ examples/miniapps/engines_cars/example/cars.py
@@ -1,9 +1,0 @@
-"""Dependency injection example, cars module."""
-
-
-class Car:
-    """Example car."""
-
-    def __init__(self, engine):
-        """Initialize instance."""
-        self._engine = engine  # Engine is injected